note/pyproject.toml
[project]
name = "note"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aerich>=0.9.2",
    "aiosqlite>=0.21.0",
    "fastapi>=0.120.4",
    "orjson>=3.11.4",
    "pika>=1.3.2",
    "python-dotenv>=1.2.1",
    "tortoise>=0.1.1",
    "tortoise-orm>=0.25.1",
    "uvicorn[standard]>=0.38.0",
]

[project.scripts]
dev = "src.scripts.main:run_server"
prod = "src.scripts.main:run_prod"
makemigrations = "src.scripts.aerich:makemigrations"
migrate = "src.scripts.aerich:migrate"
init = "src.scripts.aerich:init"
mount = "src.scripts.aerich:mount"


[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src",

]

[tool.aerich]
tortoise_orm = "src.core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
